requirements.txt
pyspark==3.5.5
requests
pyyaml
delta-spark
tenacity
typing
pathlib
python-dotenv
apache-airflow
psutil